refactor: separação de lógica
